* commit 'f759809d9d4364bc1d988e4390d3d5a33d5469e9':
  Core/PacketIO: Use new time utilities in packets instead of plain time_t/uint32

+++ src/server/game/Server/Packets/AuthenticationPackets.cpp
@@ -134,7 +134,7 @@ WorldPacket const* WorldPackets::Auth::AuthResponse::Write()
         _worldPacket << uint32(SuccessInfo->AvailableClasses->size());
         _worldPacket << uint32(SuccessInfo->Templates.size());
         _worldPacket << uint32(SuccessInfo->CurrencyID);
-        _worldPacket << int32(SuccessInfo->Time);
+        _worldPacket << SuccessInfo->Time;
 
         for (auto const& klass : *SuccessInfo->AvailableClasses)
         {
